functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.5.0
google-cloud-storage>=2.14.0
polar-sdk>=0.8.0
resend>=0.8.0
requests>=2.31.0